+++ guillotina_elasticsearch/tests/test_parser.py
@@ -1,5 +1,3 @@
-from guillotina.directives import index_field
-from guillotina.interfaces import IContainer
 from guillotina.tests import utils as test_utils
 from guillotina_elasticsearch.parser import Parser
 from guillotina_elasticsearch.tests.utils import setup_txn_on_container
@@ -12,22 +10,26 @@ import pytest
 pytestmark = [pytest.mark.asyncio]
 
 
+@pytest.mark.app_settings(
+    {
+        "applications": [
+            "guillotina",
+            "guillotina_elasticsearch",
+            "guillotina_elasticsearch.tests.test_package",
+        ]
+    }
+)
 async def test_boolean_field(es_requester):
     async with es_requester as requester:
-
-        @index_field.with_accessor(IContainer, "foo_bool", type="boolean")
-        def index_bool(obj):
-            return True
-
         container, request, txn, tm = await setup_txn_on_container(requester)  # noqa
-        params = {"depth__gte": "1", "type_name": "IItem", "foo_bool": True}
+        params = {"depth__gte": "1", "type_name": "FooItem", "item_bool": True}
         parser = Parser(None, container)
         query = parser(params)
         qq = query["query"]["bool"]["must"]
         assert "_from" not in query
-        assert qq[1]["term"]["type_name"] == "IItem"
+        assert qq[1]["term"]["type_name"] == "FooItem"
         # https://www.elastic.co/guide/en/elasticsearch/reference/current/boolean.html
-        assert qq[2]["term"]["foo_bool"] == "true"
+        assert qq[2]["term"]["item_bool"] == "true"
 
 
 @pytest.mark.app_settings(